% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backgroud_Rate.R
\name{background_rate}
\alias{background_rate}
\title{Calculate the background rate in a sequence}
\usage{
background_rate(x)
}
\arguments{
\item{x}{A vector of response values}
}
\value{
A single numeric value
}
\description{
Calculate the background rate in a sequence
}
