% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Estimation.R
\name{Est_GammaPoisson}
\alias{Est_GammaPoisson}
\title{Calculate MLE Estimates for Gamma-Poisson distribution}
\usage{
Est_GammaPoisson(x)
}
\arguments{
\item{x}{Input vector of observations}
}
\description{
Calculate MLE Estimates for Gamma-Poisson distribution
}
