% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_Estimation.R
\name{fisher_chi_sq}
\alias{fisher_chi_sq}
\title{Perform Fisher's Chi-squared test on each peptide sequence}
\usage{
fisher_chi_sq(cleaved, uncleaved)
}
\arguments{
\item{cleaved}{Vector of cleaved observatinos}

\item{uncleaved}{Vector of uncleaved observations}
}
\description{
Perform Fisher's Chi-squared test on each peptide sequence
}
