% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{full_standardize}
\alias{full_standardize}
\title{Standardize the cleaved and uncleaved as well as create the signal}
\usage{
full_standardize(df, type = "additive", scale = TRUE,
  trim_proportion = 0.25)
}
\arguments{
\item{df}{A data frame with columns cleaved and uncleaved. If the data frame is
already grouped, then all the standardization occurs within a group.}

\item{type}{Either `addititive`, `multiplicative`, or `complex`.}

\item{scale}{Should we scale the signal to have the maximum between 100 and 1000?}

\item{trim_proportion}{In the rescaling, what percent of the large values should be
removed to get to a background rate.}
}
\value{
A data frame with columns new columns cleaved_Z, uncleaved_Z, and signal. The rows correspond
        to the rows in the input data.frame.
}
\description{
There are several cases where we want to both standardize the cleaved/uncleaved
as well as calculate the signal terms
}
