% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Peaks.R
\name{identify_peaks}
\alias{identify_peaks}
\title{Identify Peaks}
\usage{
identify_peaks(x, y, method = "PoT", param = NA, min_peak_length = 1,
  merge_peak_gap = 2)
}
\arguments{
\item{x}{A vector x-values.}

\item{y}{A vector of y-values.}

\item{method}{A character string denoting which method to use. Valid options are: PeakSeg and PoT}

\item{param}{A parameter controling how many peaks are detected. For PoT, it is the treshold.}
}
\value{
A data frame with columns `Peak`, `Start`, and `End`.
}
\description{
Given (x,y) pairs of datapoints, calculate peaks.
}
