% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Peaks.R
\name{identify_peaks2}
\alias{identify_peaks2}
\title{Multivariate standardization using Cleaved/Uncleaved responses}
\usage{
identify_peaks2(data, index = "index", cleaved = "Cleaved",
  uncleaved = "Uncleaved")
}
\arguments{
\item{data}{A data frame contain an index, cleaved and uncleaved columns.}

\item{index}{Which column in the data set corresponds to the index}

\item{cleaved}{Which column in the data set corresponds to the cleaved}

\item{uncleaved}{Which column in the data set corresponds to the uncleaved values}
}
\description{
Given multiple sequences of (x,y) pairs, create a vector idenifying the peaks.
}
