% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Peptide_Peaks.R
\name{Identify_Peptide_Peaks}
\alias{Identify_Peptide_Peaks}
\title{Calculates a set of Peaks}
\usage{
Identify_Peptide_Peaks(input, output_file = "pulldown_peaks.csv",
  method = "PoT", param = NA, min_peak_length = 1,
  merge_peak_gap = 2)
}
\description{
Using the result of a pulldown_import, go through each protein and
calculate the peaks associated with each MHC.
}
\details{
This should output the data frame of peaks for each protein along with
the meta data
}
