% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulldown_Import_Metadata.R
\name{import_pulldown_metadata}
\alias{import_pulldown_metadata}
\title{Import the pulldown metadata}
\usage{
import_pulldown_metadata(file, protein_column = "protein_ID",
  position_column = "Peptide.start", meta_columns = c("library_member",
  "Peptide.seq", "Accession"))
}
\arguments{
\item{file}{The input .csv, .xls, or .xlsx file. It can be a full path to a file or a
relative path from the current working directory.}

\item{protein_column}{A character string indicating which column denotes the protein.}

\item{position_column}{A character string indicating which column corresponds to the position within a protein.}

\item{meta_columns}{A vector of character strings indicating which columns contain metadata that should be stored.}
}
\description{
Internally, the PepSeq package keeps track of peptides via the protein column and
the position within the protein.  However, that isn't consistent with other programs
so we need a way to associate the protein/position information with the library_member
code, the actual peptide sequence, as well as the Accession number
}
