% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlapping_Peaks.R
\name{Overlapping_Peaks}
\alias{Overlapping_Peaks}
\title{Calculate Overlapping Peaks}
\usage{
Overlapping_Peaks(A, B, buffer = 0, group_by = "")
}
\description{
Given two sets of peaks, figure out which peaks overlap and how much
}
\examples{
A <- data.frame(Start=c(9,20,50,100), End=c(12,25,55,199))
B <- data.frame(Start=c(3,24,60,120), End=c( 7,28,63,130))
Overlapping_Peaks(A, B)

A <- data.frame(
  Start = c( 9,20,10,25),
  End   = c(12,25,15,35),
  Grp   = c('A','A','B','B') )
B <- data.frame(
  Start = c( 3,24,10,20),
  End   = c( 7,28,13,25),
  Grp   = c('A','A','B','B') )
Overlapping_Peaks(A, B, group_by='Grp')
}
