% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulldown_Visualization.R
\name{plot_pulldown}
\alias{plot_pulldown}
\title{Create a scatterplot representing a pull-down}
\usage{
plot_pulldown(input, output_file = "pulldown.pdf", height = NULL,
  width = NULL, ymin = NULL, ymax = NULL, peaks = TRUE,
  peak_method = "PoT", peak_param = NA, scales = "free_y")
}
\arguments{
\item{input}{The input data frame read in using the
`import_pulldown()` function.}

\item{output_file}{The file to save the resulting image. In can be a full
path to the new file or a relative path from the current
working directory.}

\item{height}{The height (in inches) of the resulting file. If NULL, the
default is the number of Group-by-Treatment-by-Rep combinations
plus 3.}

\item{width}{The width (in inches) of the resulting file. If NULL, the default
is the number of sequences / 1000.}

\item{ymin}{The minimum value of the response to be shown.  If NULL, all the data is shown.}

\item{ymax}{The maximum value of the response to be shown.  If NULL, all the data is shown.}

\item{peaks}{A logical flag denoting if the peaks should be highlighted}

\item{peak_method}{Which method should be used for peak detection}

\item{peak_param}{A parameter which controls the peak finding method. For PoT, it is the threshold.}

\item{scales}{Contols if the y-scale should be 'fixed_y' across the rows or if the y-scales should be 'free_y'
to vary across the different rows.}
}
\description{
The pulldown script created by John Altin creates a .csv file
with several columns that gives the number of reads for various
peptide sequences. This script reads the .csv file and creates
a graph using ggplot2 and saves the graph to an output file. The
function also invisibly returns the dataset used to create the
graph.
}
\details{
Critically this function assumes the .csv file has columns for
`protein_ID` and `position` and that no other columns other than the read
counts start with a `X`.
}
