% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{shrink}
\alias{shrink}
\title{Remove some percentage or number of observations.}
\usage{
shrink(x, proportion = NULL, size = NULL, side = "top",
  na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{proportion}{The proportion of values to be removed. For example, if we
wanted to remove 15 percent, the proportion entered would be 0.15.}

\item{size}{The number of observations to be removed.}

\item{side}{Where should the values be removed from. \code{start} and \code{end}
remove the observations from the start and end of the vector, while
\code{top} and \code{bottom} will cause the vector to first be sorted
and then remove either the largest or smallest observations.}
}
\value{
A vector that is shorter than the input.
}
\description{
Often it is desirable to remove a certain number or percentage of a vector.
This function will accept either an input number \code{size} or percentage to be
removed.  These can be removed from either the start or end of the vector, or
the vector can be sorted first and then the items can be removed from the start
or end.
}
\examples{
set.seed(8675309)
input <- rpois(10, lambda=4)
input
shrink( input, proportion=0.1, side='start')
shrink( input, proportion=0.1, side='end')
shrink( input, proportion=0.1, side='top')
shrink( input, proportion=0.1, side='bottom')

}
