% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardize}
\alias{standardize}
\title{Standardize counts relative to uncleaved counts.}
\usage{
standardize(cleaved, uncleaved, ref = NULL, type = "additive")
}
\arguments{
\item{cleaved}{The cleaved raw counts}

\item{uncleaved}{The uncleaved raw counts}

\item{ref}{The raw counts of the reference library. Ideally these should be
identical, but the library likely isn't equally weighted}

\item{type}{Either `addititive`, `multiplicative`, or `complex`.}

\item{scale}{Should we scale to have the maximum between 100 and 1000?}
}
\description{
There are two different ways to handles standardization of
cleaved vs uncleaved. For each of the cleaved and uncleaved counts,
we first convert to the proportion of reads, and then either look at the
difference in proportion (addititive) or the ratio of the proportions
(multiplicative). In the addititive case, we might want to normalize
the differences based on the differing amounts in the reference library.
}
\examples{
df <- data.frame( cleaved   = c(20, 10,5),
                  uncleaved = c(5, 5, 2),
                  ref       = c(3, 3, 2) )
with(df, standardize( cleaved, uncleaved, ref ) )
with(df, standardize( cleaved, uncleaved, type='multiplicative' ) )
}
