% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Peaks.R
\name{identify_peaks}
\alias{identify_peaks}
\title{Identify Peaks}
\usage{
identify_peaks(x, method = "PoT", penalty = NULL, threshold = NULL)
}
\arguments{
\item{x}{A vector or data frame of observations. If a data frame, the rows should correspond
to the sequence of observations, and each column is an independent experiement.}

\item{method}{A character string denoting which method to use. Valid options are: PeakSeg and PoT}
}
\description{
Given an input data frame (or vector) of sequential observations, calculate the peaks
}
\examples{
lambdas <- rep( c(10,40,10,40,10), times=c(20,10,30,8,30) )
data <- rpois( length(lambdas), lambda = lambdas )
identify_peaks( x=data )
}
