% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulldown_Visualization.R
\name{plot_pulldown}
\alias{plot_pulldown}
\title{Create a scatterplot representing a pull-down}
\usage{
plot_pulldown(file, output_file = "pulldown.pdf", height = NULL,
  width = NULL, ymin = NULL, ymax = NULL, read_indicator = "X",
  standardization_method = "additive", protein_column = "protein_ID",
  position_column = "position", peaks = TRUE, peak_method = "PoT",
  peak_param = NA, scales = "fixed")
}
\arguments{
\item{file}{The input .csv file. It can be a full path to a file or a
relative path from the current working directory.}

\item{output_file}{The file to save the resulting image. In can be a full
path to the new file or a relative path from the current
working directory.}

\item{height}{The height (in inches) of the resulting file. If NULL, the
default is the number of Group-by-Treatment-by-Rep combinations
plus 3.}

\item{width}{The width (in inches) of the resulting file. If NULL, the default
is the number of sequences / 1000.}

\item{ymin}{The minimum value of the response to be shown.  If NULL, all the data is shown.}

\item{ymax}{The maximum value of the response to be shown.  If NULL, all the data is shown.}

\item{read_indicator}{An argument that identifies what columns are responses
to be plotted. This could be either a vector of integers
or a character string that is at the beginning of all of the
column names of the response data.}

\item{standardization_method}{The method by which the cleaved and uncleaved read
counts are combined. Valid choices are 'additive' or
'multiplicative'. The default is additive.}

\item{protein_column}{A character string indicating which column denotes the protein.}

\item{position_column}{A character string indicating which column corresponds to the position within a protein.}

\item{peaks}{A logical flag denoting if the peaks should be highlighted}

\item{peak_method}{Which method should be used for peak detection}

\item{peak_param}{A parameter which controls the peak finding method. For PoT, it is the threshold.}

\item{scales}{Contols if the y-scale should be 'fixed' across the rows or if the y-scales should be 'free'
to vary across the different rows.}
}
\description{
The pulldown script created by John Altin creates a .csv file
with several columns that gives the number of reads for various
peptide sequences. This script reads the .csv file and creates
a graph using ggplot2 and saves the graph to an output file. The
function also invisibly returns the dataset used to create the
graph.
}
\details{
Critically this function assumes the .csv file has columns for
`protein_ID` and `position` and that no other columns other than the read
counts start with a `X`.
}
