% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulldown_Visualization.R
\name{plot_pulldown}
\alias{plot_pulldown}
\title{Create a scatterplot representing a pull-down}
\usage{
plot_pulldown(input_file, output_file = "pulldown.pdf", height = NULL,
  width = NULL, standardize.method = "additive")
}
\arguments{
\item{input_file}{The input .csv file. It can be a full path to a file or a
relative path from the current working directory.}

\item{output_file}{The file to save the resulting image. In can be a full
path to the new file or a relative path from the current
working directory.}

\item{height}{The height (in inches) of the resulting file. If NULL, the
default is the number of Group-by-Treatment-by-Rep combinations
plus 3.}

\item{width}{The width (in inches) of the resulting file. If NULL, the default
is the number of sequences / 1000.}

\item{standize.method}{The method by which the cleaved and uncleaved read
counts are combined. Valid choices are 'additive' or
'multiplicative'. The default is additive.}
}
\description{
The pulldown script created by John Altin creates a .csv file
with several columns that gives the number of reads for various
peptide sequences. This script reads the .csv file and creates
a graph using ggplot2 and saves the graph to an output file. The
function also invisibly returns the dataset used to create the
graph.
}
\details{
Critically this function assumes the .csv file has columns for
`protein_ID` and `position` and that no other columns other than the read
counts start with a `X`.
}
