% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pulldown_Visualization_Shiny.R
\name{plot_pulldown_Shiny}
\alias{plot_pulldown_Shiny}
\title{Create a Shiny application showing a scatterplot representing a pull-down}
\usage{
plot_pulldown_Shiny(input_file, standardize.method = "additive",
  read_indicator = "X")
}
\arguments{
\item{input_file}{The input .csv file. It can be a full path to a file or a
relative path from the current working directory.}

\item{standardize.method}{The method by which the cleaved and uncleaved read
counts are combined. Valid choices are 'additive' or
'multiplicative'. The default is additive.}

\item{read_indicator}{What character string indicates the column is data.}
}
\description{
The pulldown script created by John Altin creates a .csv file
with several columns that gives the number of reads for various
peptide sequences. This function reads the .csv file and creates
a Shiny application that allows the user to explore the pulldown.
}
